[feat] #166 인기 펀딩 조회 로직 구현

+++ src/main/java/com/zipup/server/funding/dto/FundingSummaryResponse.java
@@ -10,7 +10,7 @@ public class FundingSummaryResponse {
   private String title;
   private String imageUrl;
   private String status;
-  private int dDay;
+  private long dDay;
   private Integer percent;
   private String organizer;
 }
